+++ authpass/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add support for permanently deleting groups and password entries. #99
 * When a conflict is detected during save, automatically merge files #80
 * Dropbox: Fixed conflict detection.
+* Google Drive: Implement rudimentary conflict detection #34
 
 # 1.8.2 - 2021-07-22